created interface for random expression generation with a schema
