+++ go/test/endtoend/vtgate/queries/random/random_test.go
@@ -18,12 +18,10 @@ package random
 
 import (
 	"fmt"
-	"math/rand"
 	"testing"
 	"time"
 
-	"vitess.io/vitess/go/vt/log"
-
+	"vitess.io/vitess/go/mysql"
 	"vitess.io/vitess/go/vt/sqlparser"
 
 	"github.com/stretchr/testify/require"
@@ -33,9 +31,6 @@ import (
 )
 
 // this test uses the AST defined in the sqlparser package to randomly generate queries
-
-type tableT = sqlparser.TableT
-type column = sqlparser.Col
 
 // if true then known failing query types are still generated by randomQuery()
 const TestFailingQueries = true
@@ -181,23 +176,23 @@ func TestRandom(t *testing.T) {
 
 	// specify the schema (that is defined in schema.sql)
 	schemaTables := []tableT{
-		{Name: sqlparser.NewTableName("emp")},
-		{Name: sqlparser.NewTableName("dept")},
+		{name: sqlparser.NewTableName("emp")},
+		{name: sqlparser.NewTableName("dept")},
 	}
-	schemaTables[0].AddColumns([]column{
-		{Name: "empno", Typ: "bigint"},
-		{Name: "ename", Typ: "varchar"},
-		{Name: "job", Typ: "varchar"},
-		{Name: "mgr", Typ: "bigint"},
-		{Name: "hiredate", Typ: "date"},
-		{Name: "sal", Typ: "bigint"},
-		{Name: "comm", Typ: "bigint"},
-		{Name: "deptno", Typ: "bigint"},
+	schemaTables[0].addColumns([]column{
+		{name: "empno", typ: "bigint"},
+		{name: "ename", typ: "varchar"},
+		{name: "job", typ: "varchar"},
+		{name: "mgr", typ: "bigint"},
+		{name: "hiredate", typ: "date"},
+		{name: "sal", typ: "bigint"},
+		{name: "comm", typ: "bigint"},
+		{name: "deptno", typ: "bigint"},
 	}...)
-	schemaTables[1].AddColumns([]column{
-		{Name: "deptno", Typ: "bigint"},
-		{Name: "dname", Typ: "varchar"},
-		{Name: "loc", Typ: "varchar"},
+	schemaTables[1].addColumns([]column{
+		{name: "deptno", typ: "bigint"},
+		{name: "dname", typ: "varchar"},
+		{name: "loc", typ: "varchar"},
 	}...)
 
 	endBy := time.Now().Add(10 * time.Second)
@@ -206,6 +201,9 @@ func TestRandom(t *testing.T) {
 	for time.Now().Before(endBy) && (!t.Failed() || TestFailingQueries) {
 		query := sqlparser.String(randomQuery(schemaTables, 3, 3))
 		_, vtErr := mcmp.ExecAllowAndCompareError(query)
+		if sqlError, ok := vtErr.(*mysql.SQLError); ok && sqlError.Message == "EOF" {
+			break
+		}
 		// this assumes all queries are valid mysql queries
 		if vtErr != nil {
 			fmt.Println(query)
@@ -217,301 +215,4 @@ func TestRandom(t *testing.T) {
 		queryCount++
 	}
 	fmt.Printf("Queries successfully executed: %d\n", queryCount)
-}
-
-// TODO: bunch of TestFailingQueries checks were deleted by refactor to use AST
-func randomQuery(schemaTables []tableT, maxAggrs, maxGroupBy int) *sqlparser.Select {
-	sel := &sqlparser.Select{}
-	sel.SetComments(sqlparser.Comments{"/*vt+ PLANNER=Gen4 */"})
-
-	// select distinct (fails with group by bigint)
-	isDistinct := rand.Intn(2) < 1
-	if isDistinct {
-		sel.MakeDistinct()
-	}
-
-	// create both tables and join at the same time since both occupy the from clause
-	tables, isJoin := createTablesAndJoin(schemaTables, sel)
-
-	groupExprs, groupSelectExprs, grouping := createGroupBy(tables, maxGroupBy)
-	sel.AddSelectExprs(groupSelectExprs)
-	sel.GroupBy = groupExprs
-	aggrExprs, aggregates := createAggregations(tables, maxAggrs)
-	sel.AddSelectExprs(aggrExprs)
-
-	// can add both aggregate and grouping columns to order by
-	isOrdered := rand.Intn(2) < 1
-	if isOrdered && (!isDistinct || TestFailingQueries) && (!isJoin || TestFailingQueries) {
-		addOrderBy(sel)
-	}
-
-	// where
-	sel.AddWhere(sqlparser.AndExpressions(createWherePredicates(tables, false)...))
-
-	// random predicate expression
-	if rand.Intn(2) < 1 {
-		predRandomExpr, _ := getRandomExpr(tables)
-		sel.AddWhere(predRandomExpr)
-	}
-
-	// having
-	sel.AddHaving(sqlparser.AndExpressions(createHavingPredicates(tables)...))
-	if rand.Intn(2) < 1 {
-		sel.AddHaving(sqlparser.AndExpressions(createWherePredicates(tables, false)...))
-	}
-
-	// only add a limit if the grouping columns are ordered
-	if rand.Intn(2) < 1 && isOrdered {
-		sel.Limit = createLimit()
-	}
-
-	var newTable tableT
-	// add random expression to select
-	isRandomExpr := rand.Intn(2) < 1
-	randomExpr, typ := getRandomExpr(tables)
-	if isRandomExpr && (!isDistinct || TestFailingQueries) && (!isJoin || TestFailingQueries) {
-		sel.SelectExprs = append(sel.SelectExprs, sqlparser.NewAliasedExpr(randomExpr, "crandom0"))
-		newTable.AddColumns(column{
-			Name: "crandom0",
-			Typ:  typ,
-		})
-	}
-
-	// alias grouping and aggregate columns
-	for i := range grouping {
-		grouping[i].Name = fmt.Sprintf("cgroup%d", i)
-	}
-	for i := range aggregates {
-		aggregates[i].Name = fmt.Sprintf("cgroup%d", i)
-	}
-
-	// add them to newTable
-	newTable.AddColumns(grouping...)
-	newTable.AddColumns(aggregates...)
-
-	// add new table to schemaTables
-	newTable.Name = sqlparser.NewDerivedTable(false, sel)
-	schemaTables = append(schemaTables, newTable)
-
-	// derived tables (partially unsupported)
-	if rand.Intn(10) < 1 && TestFailingQueries {
-		sel = randomQuery(schemaTables, 3, 3)
-	}
-
-	return sel
-}
-
-func createTablesAndJoin(schemaTables []tableT, sel *sqlparser.Select) ([]tableT, bool) {
-	var tables []tableT
-	// add at least one of original emp/dept tables for now because derived tables have nil columns
-	tables = append(tables, schemaTables[rand.Intn(2)])
-
-	sel.From = append(sel.From, newAliasedTable(tables[0], "tbl0"))
-	tables[0].SetName("tbl0")
-
-	numTables := rand.Intn(len(schemaTables))
-	for i := 0; i < numTables; i++ {
-		tables = append(tables, randomEl(schemaTables))
-		sel.From = append(sel.From, newAliasedTable(tables[i+1], fmt.Sprintf("tbl%d", i+1)))
-		tables[i+1].SetName(fmt.Sprintf("tbl%d", i+1))
-	}
-
-	isJoin := rand.Intn(2) < 1
-	if isJoin {
-		newTable := randomEl(schemaTables)
-		tables = append(tables, newTable)
-
-		// create the join before aliasing
-		newJoinTableExpr := createJoin(tables, sel)
-
-		// alias
-		tables[numTables+1].SetName(fmt.Sprintf("tbl%d", numTables+1))
-
-		// create the condition after aliasing
-		newJoinTableExpr.Condition = sqlparser.NewJoinCondition(sqlparser.AndExpressions(createWherePredicates(tables, true)...), nil)
-		sel.From[numTables] = newJoinTableExpr
-	}
-
-	return tables, isJoin
-}
-
-// creates a left join (without the condition) between the last table in sel and newTable
-// tables should have one more table than sel
-func createJoin(tables []tableT, sel *sqlparser.Select) *sqlparser.JoinTableExpr {
-	n := len(sel.From)
-	if len(tables) != n+1 {
-		log.Fatalf("sel has %d tables and tables has %d tables", len(sel.From), n)
-	}
-
-	return sqlparser.NewJoinTableExpr(sel.From[n-1], sqlparser.LeftJoinType, newAliasedTable(tables[n], fmt.Sprintf("tbl%d", n)), nil)
-}
-
-// adds grouping columns to sel.GroupBy and optionally to sel.SelectExprs and sel.OrderBy
-// TODO: maybe change to return this stuff instead
-func createGroupBy(tables []tableT, maxGB int) (groupBy sqlparser.GroupBy, groupSelectExprs sqlparser.SelectExprs, grouping []column) {
-	numGBs := rand.Intn(maxGB)
-	for i := 0; i < numGBs; i++ {
-		tblIdx := rand.Intn(len(tables))
-		col := randomEl(tables[tblIdx].Cols)
-		groupBy = append(groupBy, newColumn(col))
-
-		// add to select
-		if rand.Intn(2) < 1 {
-			groupSelectExprs = append(groupSelectExprs, newAliasedColumn(col, fmt.Sprintf("cgroup%d", i)))
-			grouping = append(grouping, col)
-		}
-	}
-
-	return groupBy, groupSelectExprs, grouping
-}
-
-// adds aggregation columns to sel.SelectExprs and optionally to sel.OrderBy
-// TODO: maybe change to return this stuff instead
-func createAggregations(tables []tableT, maxAggrs int) (aggrExprs sqlparser.SelectExprs, aggregates []column) {
-	aggregations := []func(col column) sqlparser.Expr{
-		func(_ column) sqlparser.Expr { return &sqlparser.CountStar{} },
-		func(col column) sqlparser.Expr { return &sqlparser.Count{Args: sqlparser.Exprs{newColumn(col)}} },
-		func(col column) sqlparser.Expr { return &sqlparser.Sum{Arg: newColumn(col)} },
-		// func(col column) sqlparser.Expr { return &sqlparser.Avg{Arg: newAggregateExpr(col)} },
-		func(col column) sqlparser.Expr { return &sqlparser.Min{Arg: newColumn(col)} },
-		func(col column) sqlparser.Expr { return &sqlparser.Max{Arg: newColumn(col)} },
-	}
-
-	numAggrs := rand.Intn(maxAggrs) + 1
-	for i := 0; i < numAggrs; i++ {
-		tblIdx, aggrIdx := rand.Intn(len(tables)), rand.Intn(len(aggregations))
-		col := randomEl(tables[tblIdx].Cols)
-		newAggregate := aggregations[aggrIdx](col)
-		aggrExprs = append(aggrExprs, sqlparser.NewAliasedExpr(newAggregate, fmt.Sprintf("caggr%d", i)))
-
-		if aggrIdx <= 1 /* CountStar and Count */ {
-			col.Typ = "bigint"
-		} else if _, ok := newAggregate.(*sqlparser.Avg); ok && col.GetColumnName() == "bigint" {
-			col.Typ = "decimal"
-		}
-
-		col.Name = sqlparser.String(newAggregate)
-		aggregates = append(aggregates, col)
-	}
-	return aggrExprs, aggregates
-}
-
-// orders on all non-aggregate SelectExprs and independently at random on all aggregate SelectExprs of sel
-func addOrderBy(sel *sqlparser.Select) {
-	for _, selExpr := range sel.SelectExprs {
-		if aliasedExpr, ok := selExpr.(*sqlparser.AliasedExpr); ok {
-			// if the SelectExpr is non-aggregate (the AliasedExpr has Expr of type ColName)
-			// then add to the order by
-			if colName, ok1 := aliasedExpr.Expr.(*sqlparser.ColName); ok1 {
-				sel.AddOrder(sqlparser.NewOrder(colName, getRandomOrderDirection()))
-			} else if rand.Intn(2) < 1 {
-				sel.AddOrder(sqlparser.NewOrder(aliasedExpr.Expr, getRandomOrderDirection()))
-			}
-		}
-	}
-}
-
-// returns the predicates as a sqlparser.Exprs (slice of sqlparser.Expr's)
-func createWherePredicates(tables []tableT, isJoin bool) (predicates sqlparser.Exprs) {
-	// if creating predicates for a join,
-	// then make sure predicates are created for the last two tables (which are being joined)
-	incr := 0
-	if isJoin && len(tables) > 2 {
-		incr += len(tables) - 2
-	}
-
-	for idx1 := range tables {
-		for idx2 := range tables {
-			// fmt.Printf("predicate tables:\n%v\n idx1: %d idx2: %d, incr: %d", tables, idx1, idx2, incr)
-			if idx1 >= idx2 || idx1 < incr || idx2 < incr {
-				continue
-			}
-			noOfPredicates := rand.Intn(2)
-			if isJoin {
-				noOfPredicates++
-			}
-
-			for i := 0; noOfPredicates > 0; i++ {
-				col1 := randomEl(tables[idx1].Cols)
-				col2 := randomEl(tables[idx2].Cols)
-
-				// prevent infinite loops
-				if i > 50 {
-					predicates = append(predicates, sqlparser.NewComparisonExpr(getRandomComparisonExprOperator(), newColumn(col1), newColumn(col2), nil))
-					break
-				}
-
-				if col1.Typ != col2.Typ {
-					continue
-				}
-
-				predicates = append(predicates, sqlparser.NewComparisonExpr(getRandomComparisonExprOperator(), newColumn(col1), newColumn(col2), nil))
-				noOfPredicates--
-			}
-		}
-	}
-
-	// make sure the join predicate is never empty
-	if len(predicates) == 0 && isJoin {
-		predRandomExpr, _ := getRandomExpr(tables)
-		predicates = append(predicates, predRandomExpr)
-	}
-
-	return predicates
-}
-
-func createHavingPredicates(tables []tableT) (havingPredicates sqlparser.Exprs) {
-	aggrSelectExprs, _ := createAggregations(tables, 3)
-	for i := range aggrSelectExprs {
-		if aliasedExpr, ok := aggrSelectExprs[i].(*sqlparser.AliasedExpr); ok {
-			predRandomExpr, _ := getRandomExpr(tables)
-			havingPredicates = append(havingPredicates, sqlparser.NewComparisonExpr(getRandomComparisonExprOperator(), aliasedExpr.Expr, predRandomExpr, nil))
-		}
-	}
-	return havingPredicates
-}
-
-// creates sel.Limit
-func createLimit() *sqlparser.Limit {
-	limitNum := rand.Intn(10)
-	if rand.Intn(2) < 1 {
-		offset := rand.Intn(10)
-		return sqlparser.NewLimit(offset, limitNum)
-	} else {
-		return sqlparser.NewLimitWithoutOffset(limitNum)
-	}
-
-}
-
-// returns a random expression and its type
-func getRandomExpr(tables []tableT) (sqlparser.Expr, string) {
-	seed := time.Now().UnixNano()
-	g := sqlparser.NewGenerator(seed, 2, tables...)
-	return g.Expression()
-}
-
-func newAliasedTable(tbl tableT, alias string) *sqlparser.AliasedTableExpr {
-	return sqlparser.NewAliasedTableExpr(tbl.Name, alias)
-}
-
-func newAliasedColumn(col column, alias string) *sqlparser.AliasedExpr {
-	return sqlparser.NewAliasedExpr(newColumn(col), alias)
-}
-
-func newColumn(col column) *sqlparser.ColName {
-	return sqlparser.NewColNameWithQualifier(col.Name, sqlparser.NewTableName(col.TableName))
-}
-
-func getRandomComparisonExprOperator() sqlparser.ComparisonExprOperator {
-	// =, <, >, <=, >=, !=, <=>
-	return randomEl([]sqlparser.ComparisonExprOperator{0, 1, 2, 3, 4, 5, 6})
-}
-
-func getRandomOrderDirection() sqlparser.OrderDirection {
-	// asc, desc
-	return randomEl([]sqlparser.OrderDirection{0, 1})
-}
-
-func randomEl[K any](in []K) K {
-	return in[rand.Intn(len(in))]
 }
